+++ tests/unit/stats/test_aggregation.py
@@ -12,103 +12,97 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import mock
 import unittest
 
 from opencensus.stats import aggregation as aggregation_module
-
-
-class TestBaseAggregation(unittest.TestCase):
-    def test_constructor_defaults(self):
-        base_aggregation = aggregation_module.BaseAggregation()
-
-        self.assertEqual(aggregation_module.Type.NONE,
-                         base_aggregation.aggregation_type)
-        self.assertEqual([], base_aggregation.buckets)
-
-    def test_constructor_explicit(self):
-
-        buckets = ["test"]
-        base_aggregation = aggregation_module.BaseAggregation(buckets=buckets)
-
-        self.assertEqual(aggregation_module.Type.NONE,
-                         base_aggregation.aggregation_type)
-        self.assertEqual(["test"], base_aggregation.buckets)
+from opencensus.stats import measure as measure_module
+from opencensus.metrics.export import value
 
 
 class TestSumAggregation(unittest.TestCase):
-    def test_constructor_defaults(self):
+    def test_new_aggregation_data_defaults(self):
+        measure = mock.Mock(spec=measure_module.MeasureInt)
         sum_aggregation = aggregation_module.SumAggregation()
+        agg_data = sum_aggregation.new_aggregation_data(measure)
+        self.assertEqual(0, agg_data.sum_data)
+        self.assertEqual(value.ValueLong, agg_data.value_type)
 
-        self.assertEqual(0, sum_aggregation.sum.sum_data)
-        self.assertEqual(aggregation_module.Type.SUM,
-                         sum_aggregation.aggregation_type)
+    def test_new_aggregation_data_explicit(self):
+        measure = mock.Mock(spec=measure_module.MeasureInt)
+        sum_aggregation = aggregation_module.SumAggregation(sum=1)
+        agg_data = sum_aggregation.new_aggregation_data(measure)
+        self.assertEqual(1, agg_data.sum_data)
+        self.assertEqual(value.ValueLong, agg_data.value_type)
 
-    def test_constructor_explicit(self):
-        sum = 1
+    def test_new_aggregation_data_float(self):
+        measure = mock.Mock(spec=measure_module.MeasureFloat)
+        sum_aggregation = aggregation_module.SumAggregation()
+        agg_data = sum_aggregation.new_aggregation_data(measure)
+        self.assertEqual(0, agg_data.sum_data)
+        self.assertEqual(value.ValueDouble, agg_data.value_type)
 
-        sum_aggregation = aggregation_module.SumAggregation(sum=sum)
-
-        self.assertEqual(1, sum_aggregation.sum.sum_data)
-        self.assertEqual(aggregation_module.Type.SUM,
-                         sum_aggregation.aggregation_type)
+    def test_new_aggregation_data_bad(self):
+        measure = mock.Mock(spec=measure_module.BaseMeasure)
+        sum_aggregation = aggregation_module.SumAggregation()
+        with self.assertRaises(ValueError):
+            sum_aggregation.new_aggregation_data(measure)
 
 
 class TestCountAggregation(unittest.TestCase):
-    def test_constructor_defaults(self):
+    def test_new_aggregation_data_defaults(self):
         count_aggregation = aggregation_module.CountAggregation()
+        agg_data = count_aggregation.new_aggregation_data()
+        self.assertEqual(0, agg_data.count_data)
 
-        self.assertEqual(0, count_aggregation.count.count_data)
-        self.assertEqual(aggregation_module.Type.COUNT,
-                         count_aggregation.aggregation_type)
-
-    def test_constructor_explicit(self):
-        count = 4
-
-        count_aggregation = aggregation_module.CountAggregation(count=count)
-
-        self.assertEqual(4, count_aggregation.count.count_data)
-        self.assertEqual(aggregation_module.Type.COUNT,
-                         count_aggregation.aggregation_type)
+    def test_new_aggregation_data_explicit(self):
+        count_aggregation = aggregation_module.CountAggregation(count=4)
+        agg_data = count_aggregation.new_aggregation_data()
+        self.assertEqual(4, agg_data.count_data)
 
 
 class TestLastValueAggregation(unittest.TestCase):
-    def test_constructor_defaults(self):
+    def test_new_aggregation_data_defaults(self):
+        measure = mock.Mock(spec=measure_module.MeasureInt)
         last_value_aggregation = aggregation_module.LastValueAggregation()
+        agg_data = last_value_aggregation.new_aggregation_data(measure)
+        self.assertEqual(0, agg_data.value)
+        self.assertEqual(value.ValueLong, agg_data.value_type)
 
-        self.assertEqual(0, last_value_aggregation.value)
-        self.assertEqual(aggregation_module.Type.LASTVALUE,
-                         last_value_aggregation.aggregation_type)
-
-    def test_constructor_explicit(self):
-        val = 16
+    def test_new_aggregation_data_explicit(self):
+        measure = mock.Mock(spec=measure_module.MeasureInt)
         last_value_aggregation = aggregation_module.LastValueAggregation(
-            value=val)
+            value=6)
+        agg_data = last_value_aggregation.new_aggregation_data(measure)
+        self.assertEqual(6, agg_data.value)
+        self.assertEqual(value.ValueLong, agg_data.value_type)
 
-        self.assertEqual(16, last_value_aggregation.value)
-        self.assertEqual(aggregation_module.Type.LASTVALUE,
-                         last_value_aggregation.aggregation_type)
+    def test_new_aggregation_data_float(self):
+        measure = mock.Mock(spec=measure_module.MeasureFloat)
+        last_value_aggregation = aggregation_module.LastValueAggregation()
+        agg_data = last_value_aggregation.new_aggregation_data(measure)
+        self.assertEqual(0, agg_data.value)
+        self.assertEqual(value.ValueDouble, agg_data.value_type)
+
+    def test_new_aggregation_data_bad(self):
+        measure = mock.Mock(spec=measure_module.BaseMeasure)
+        last_value_aggregation = aggregation_module.LastValueAggregation()
+        with self.assertRaises(ValueError):
+            last_value_aggregation.new_aggregation_data(measure)
 
 
 class TestDistributionAggregation(unittest.TestCase):
-    def test_constructor_defaults(self):
+    def test_new_aggregation_data_defaults(self):
         distribution_aggregation = aggregation_module.DistributionAggregation()
+        agg_data = distribution_aggregation.new_aggregation_data()
+        self.assertEqual([], agg_data.bounds)
 
-        self.assertEqual([], distribution_aggregation.boundaries.boundaries)
-        self.assertEqual({}, distribution_aggregation.distribution)
-        self.assertEqual(aggregation_module.Type.DISTRIBUTION,
-                         distribution_aggregation.aggregation_type)
-
-    def test_constructor_explicit(self):
+    def test_new_aggregation_data_explicit(self):
         boundaries = [1, 2]
-        distribution = [0, 1, 2]
         distribution_aggregation = aggregation_module.DistributionAggregation(
-            boundaries=boundaries, distribution=distribution)
-
-        self.assertEqual([1, 2],
-                         distribution_aggregation.boundaries.boundaries)
-        self.assertEqual([0, 1, 2], distribution_aggregation.distribution)
-        self.assertEqual(aggregation_module.Type.DISTRIBUTION,
-                         distribution_aggregation.aggregation_type)
+            boundaries=boundaries)
+        agg_data = distribution_aggregation.new_aggregation_data()
+        self.assertEqual(boundaries, agg_data.bounds)
 
     def test_init_bad_boundaries(self):
         """Check that boundaries must be sorted and unique."""
@@ -120,9 +114,7 @@ class TestDistributionAggregation(unittest.TestCase):
     def test_init_negative_boundaries(self):
         """Check that non-positive boundaries are dropped."""
         da = aggregation_module.DistributionAggregation([-2, -1, 0, 1, 2])
-        self.assertEqual(da.boundaries.boundaries, [1, 2])
-        self.assertEqual(da.aggregation_data.bounds, [1, 2])
+        self.assertEqual(da.new_aggregation_data().bounds, [1, 2])
 
         da2 = aggregation_module.DistributionAggregation([-2, -1])
-        self.assertEqual(da2.boundaries.boundaries, [])
-        self.assertEqual(da2.aggregation_data.bounds, [])
+        self.assertEqual(da2.new_aggregation_data().bounds, [])